refactor: move file_search to top-level
